Update TensorRT-LLM (#787)
* Update TensorRT-LLM

---------

+++ examples/gptneox/weight.py
@@ -1,4 +1,4 @@
-# SPDX-FileCopyrightText: Copyright (c) 2022-2023 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-FileCopyrightText: Copyright (c) 2022-2024 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -59,7 +59,7 @@ def preprocess_groupwise_weight_params(qweight_unpacked_int8, scales_fp16,
     preprocessor = torch.ops.fastertransformer.preprocess_weights_for_mixed_gemm
 
     qweight_interleaved = preprocessor(packer(qweight_unpacked_int8),
-                                       torch.quint4x2).view(torch.int8)
+                                       torch.quint4x2).view(torch.float16)
 
     # zeros = zeros * scales
     zeros_x_scales_fp16 = (-qzeros_unpacked_int8 + 8 * UINT4_TO_INT4_FLAG -
